+++ R/dock_from_renv.R
@@ -22,7 +22,7 @@ available_distros <- c(
 #' @param repos character. The URL(s) of the repositories to use for `options("repos")`.
 #' @param extra_sysreqs character vector. Extra debian system requirements.
 #'    Will be installed with apt-get install.
-#' @param keep_renv_version boolean. If `TRUE` the version of renv in the lockfile
+#' @param fix_renv_version boolean. If `TRUE` the version of renv in the lockfile
 #'    will be used for the `renv::restore()` command
 #' @importFrom utils getFromNamespace
 #' @return A R6 object of class `Dockerfile`.
@@ -55,7 +55,7 @@ dock_from_renv <- function(
   repos = c(CRAN = "https://cran.rstudio.com/"),
   expand = FALSE,
   extra_sysreqs = NULL,
-  keep_renv_version = FALSE
+  fix_renv_version = FALSE
 ) {
   distro <- match.arg(distro, available_distros)
 
@@ -241,8 +241,8 @@ dock_from_renv <- function(
     )
   )
 
-  # check if keep_renv_version is true
-  if (keep_renv_version){
+  # check if fix_renv_version is true
+  if (fix_renv_version){
     dock$RUN("R -e 'install.packages(\"remotes\")'")
       install_renv_string <- paste0(
         "R -e 'remotes::install_version(\"renv\", version = ",