+++ R/dock_from_renv.R
@@ -18,27 +18,26 @@ pkg_system_requirements_mem <- memoise::memoise(
 #' Create a Dockerfile from an `renv.lock` file
 #'
 #' @param lockfile Path to an `renv.lock` file to use as an input..
-#' @param FROM Docker image to start FROM Default is
-#'     FROM rocker/r-base
-#' @param AS The AS of the Dockerfile. Default it NULL.
-#' @param distro One of "focal", "bionic", "xenial", "centos7",
-#'     or "centos8". See available distributions
-#'     at https://hub.docker.com/r/rstudio/r-base/.
-#' @param sysreqs boolean. If `TRUE`, the Dockerfile
-#'     will contain sysreq installation.
+#' @param FROM Docker image to start FROM Default is FROM rocker/r-base
+#' @param AS The AS of the Dockerfile. Default it `NULL`.
+#' @param distro One of "focal", "bionic", "xenial", "centos7",or "centos8". See available distributions     at https://hub.docker.com/r/rstudio/r-base/.
+#' @param sysreqs boolean. If `TRUE`, the Dockerfile will contain sysreq installation.
 #' @param expand boolean. If `TRUE` each system requirement will have its own `RUN` line.
 #' @param repos character. The URL(s) of the repositories to use for `options("repos")`.
 #' @param extra_sysreqs character vector. Extra debian system requirements.
 #'    Will be installed with apt-get install.
-#' @param renv_version character. The {renv} version to use in the generated Dockerfile. 
-#' By default, it is set to the version specified in the `renv.lock` file. 
-#' If the `renv.lock` file does not specify a {renv} version, the version of {renv} bundled with {dockerfiler}, specifically `dockerfiler::renv$the$metadata$version`, will be used. If you set it to NULL, the latest available version of {renv} will be used.
+#' @param renv_version character. The renv version to use in the generated Dockerfile. By default, it is set to the version specified in the `renv.lock` file. 
+#'   If the `renv.lock` file does not specify a renv version,
+#'   the version of renv bundled with dockerfiler, 
+#'   specifically `r dockerfiler::renv$initialize();toString(dockerfiler::renv$the$metadata$version)`, will be used.
+#'   If you set it to `NULL`, the latest available version of renv will be used.
+#' @param use_pak boolean. If `TRUE` use pak to deal with dependencies  during `renv::restore()`. FALSE by default
 #' @importFrom utils getFromNamespace
 #' @return A R6 object of class `Dockerfile`.
 #' @details
 #'
 #' System requirements for packages are provided
-#' through RStudio Package Manager via the {pak}
+#' through RStudio Package Manager via the pak
 #' package. The install commands provided from pak
 #' are added as `RUN` directives within the `Dockerfile`.
 #'
@@ -65,6 +64,7 @@ dock_from_renv <- function(
   repos = c(CRAN = "https://cran.rstudio.com/"),
   expand = FALSE,
   extra_sysreqs = NULL,
+  use_pak = FALSE,
   renv_version
 ) {
   distro <- match.arg(distro, available_distros)
@@ -256,7 +256,8 @@ dock_from_renv <- function(
 
   dock$RUN(
     sprintf(
-      "echo \"options(renv.config.pak.enabled = TRUE, repos = %s, download.file.method = 'libcurl', Ncpus = 4)\" | tee /usr/local/lib/R/etc/Rprofile.site | tee /usr/lib/R/etc/Rprofile.site",
+      "echo \"options(renv.config.pak.enabled = %s, repos = %s, download.file.method = 'libcurl', Ncpus = 4)\" | tee /usr/local/lib/R/etc/Rprofile.site | tee /usr/lib/R/etc/Rprofile.site",
+      use_pak,
       repos_as_character
     )
   )