+++ R/renv.R
@@ -5,13 +5,12 @@
 #
 
 
-#' Internalised {renv}
+#' Internalised renv
 #'
 #' https://rstudio.github.io/renv/reference/vendor.html?q=vendor
 #'
 #' @export
 #' @name renv
-#' @alias renv
 #' @docType data
 renv <- new.env(parent = new.env())